best fit rectangle method for image analysis
The best-fit rectangle approach is a combination of the Ferret method and the least 2nd moments minimization, only requiring calculation of three moments about the center of gravity, and maximum and minimum co-ordinates in a co-ordinate system oriented in the direction of the axis of the least 2nd moments, and a simple area ratio. It is a simple rotation-invariance method, reflecting shape (elongation and angularity).